rest: improve RestClient API / add key ctor
This CL cleans up RestClient API, fixing couple of incosistencies and
shortcommings, like:

  - when (*RestClient).Options is changed by the user, changes are
    not reflected in *Auth
  - all the tests were modyfing global client instance, not it's more
    COW like
  - RestClient and ClientOptions are fully zero-value friendly
  - ClientOptions properly manages TLS / NoTLS transitions

RestClient would require more work, like support for token auth
(it's using BasicAuth for requests), proper TLS/NoTLS testing. This is
what I've seen from the code itself, possibly there's more if compared
to SPEC.md.

However with this implementation I'm able to implement RealtimeClient,
which is ignored from compilation via build tags, as there's not sense
to refactor it, as it's going to be rewritten.

Changes in Go API:

- added function for initializing RestClient from key:

  ```
  client, err := ably.NewRestClient("xxx:xxx")
  ```

- RestClient can be initialized from ClientOptions for more control
  over behaviour: (added correct handling of zero-value struct)

  ```
  opts := &ably.ClientOptions{Token: "xxx", Secret: "xxx"}
  client := &alby.RestClient{Options: opts}
  ```

- ClientOptions has no `Key` nor `NoTLS` fields, as it's not possible to
  keep them in sync with dependant fields like Token, Secret for Key or
  RestEndpoint, RealtimeEndpoint for `NoTLS`, e.g.

  ```go
  /* old API */
  client := ably.NewRestClient(&ably.Options{Key: "xxx:xxx"})
  // ...
  client.Options.Key = "yyy:yyy"
  // all calls to REST API are going to fail, as Token and Secret fields
  // are still "xxx"

  /* new API */
  client, err := ably.NewRestClient("xxx:xxx")
  // ...
  err := client.SetKey("yyy:yyy")
  // correctly handled here
  ```

The same applies to NoTLS, where flipping its value should rewrite URLs
of RestEndpoint and RealtimeEndpoint. With new API it's done with:

  ```go
  client.SetTLS(false)
  ```